Refactor basis module structure
