+++ src/aspire/basis/ffb_2d.py
@@ -6,7 +6,7 @@ from scipy.fftpack import fft, ifft
 from scipy.special import jv
 
 from aspire.basis.basis_utils import lgwt
-from aspire.basis.fb_2d import FBBasis2D
+from aspire.basis import FBBasis2D
 from aspire.image import Image
 from aspire.nufft import anufft, nufft
 from aspire.utils.matlab_compat import m_reshape
