+++ src/aspire/commands/cov3d.py
@@ -2,7 +2,7 @@ import logging
 
 import click
 
-from aspire.basis.fb_3d import FBBasis3D
+from aspire.basis import FBBasis3D
 from aspire.estimation.covar import CovarianceEstimator
 from aspire.estimation.mean import MeanEstimator
 from aspire.estimation.noise import WhiteNoiseEstimator
