+++ src/aspire/commands/denoise.py
@@ -2,7 +2,7 @@ import logging
 
 import click
 
-from aspire.basis.ffb_2d import FFBBasis2D
+from aspire.basis import FFBBasis2D
 from aspire.denoising.denoiser_cov2d import DenoiserCov2D
 from aspire.estimation.noise import AnisotropicNoiseEstimator, WhiteNoiseEstimator
 from aspire.source.relion import RelionSource
