+++ tests/test_Diracbasis.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 
 import numpy as np
 
-from aspire.basis.dirac import DiracBasis
+from aspire.basis import DiracBasis
 from aspire.utils.matlab_compat import m_flatten
