+++ tests/test_FPSWFbasis2D.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 
 import numpy as np
 
-from aspire.basis.fpswf_2d import FPSWFBasis2D
+from aspire.basis import FPSWFBasis2D
 
 DATA_DIR = os.path.join(os.path.dirname(__file__), "saved_test_data")
