+++ tests/test_PolarBasis2D.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 
 import numpy as np
 
-from aspire.basis.polar_2d import PolarBasis2D
+from aspire.basis import PolarBasis2D
 from aspire.image import Image
 from aspire.utils.matlab_compat import m_reshape
 from aspire.utils.types import complex_type, utest_tolerance
