+++ tests/test_covar2d.py
@@ -4,7 +4,7 @@ from unittest import TestCase
 
 import numpy as np
 
-from aspire.basis.ffb_2d import FFBBasis2D
+from aspire.basis import FFBBasis2D
 from aspire.estimation.covar2d import RotCov2D
 from aspire.source.simulation import Simulation
 from aspire.utils.filters import RadialCTFFilter, ScalarFilter
