+++ tutorials/examples/cov2d_experiment.py.dontrun
@@ -4,7 +4,7 @@ This script illustrates denoising 2D images using batched Cov2D class
 from experime dataset and outputing to mrcs file.
 """
 
-from aspire.basis.ffb_2d import FFBBasis2D
+from aspire.basis import FFBBasis2D
 from aspire.denoising.denoiser_cov2d import DenoiserCov2D
 from aspire.estimation.noise import AnisotropicNoiseEstimator
 from aspire.source.relion import RelionSource
