+++ tutorials/examples/cov2d_simulation.py
@@ -12,7 +12,7 @@ import matplotlib.pyplot as plt
 import mrcfile
 import numpy as np
 
-from aspire.basis.ffb_2d import FFBBasis2D
+from aspire.basis import FFBBasis2D
 from aspire.estimation.covar2d import RotCov2D
 from aspire.source.simulation import Simulation
 from aspire.utils.filters import RadialCTFFilter, ScalarFilter
