+++ tutorials/examples/image_expansion.py
@@ -10,10 +10,10 @@ import timeit
 import matplotlib.pyplot as plt
 import numpy as np
 
-from aspire.basis.fb_2d import FBBasis2D
-from aspire.basis.ffb_2d import FFBBasis2D
-from aspire.basis.fpswf_2d import FPSWFBasis2D
-from aspire.basis.pswf_2d import PSWFBasis2D
+from aspire.basis import FBBasis2D
+from aspire.basis import FFBBasis2D
+from aspire.basis import FPSWFBasis2D
+from aspire.basis import PSWFBasis2D
 from aspire.utils.matrix import anorm
 
 logger = logging.getLogger(__name__)
